backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "UnwindDidital Task"
authors = ["1koro <ok@hustn.cn>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"
loguru = "^0.6.0"
peewee-async = "^0.8.0"
asyncpg = "^0.26.0"
aiosql = "^6.1"
fastapi-utils = "^0.2.1"
gspread = "^5.4.0"
requests-async = "^0.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
